Fixing and refactoring MultiYtClient - codestyle
