+++ yt/java/ytsaurus-client/src/main/java/tech/ytsaurus/client/MultiYTsaurusClient.java
@@ -3,10 +3,15 @@ package tech.ytsaurus.client;
 import java.io.Closeable;
 import java.io.IOException;
 import java.time.Duration;
+import java.time.Instant;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Function;
 import java.util.function.Supplier;
+
+import javax.annotation.Nullable;
 
 import tech.ytsaurus.client.request.AbstractLookupRowsRequest;
 import tech.ytsaurus.client.request.MultiLookupRowsRequest;
@@ -125,8 +130,7 @@ public class MultiYTsaurusClient implements ImmutableTransactionalClient, Closea
     public <T> CompletableFuture<List<T>> lookupRows(
             AbstractLookupRowsRequest<?, ?> request,
             YTreeRowSerializer<T> serializer
-    )
-    {
+    ) {
         return executor.execute((client) -> client.lookupRows(request, serializer));
     }
 
@@ -139,8 +143,7 @@ public class MultiYTsaurusClient implements ImmutableTransactionalClient, Closea
     public <T> CompletableFuture<List<List<T>>> multiLookupRows(
             MultiLookupRowsRequest request,
             YTreeRowSerializer<T> serializer
-    )
-    {
+    ) {
         return executor.execute((client) -> client.multiLookupRows(request, serializer));
     }
 
@@ -163,15 +166,13 @@ public class MultiYTsaurusClient implements ImmutableTransactionalClient, Closea
     public <T> CompletableFuture<List<T>> selectRows(
             SelectRowsRequest request,
             YTreeRowSerializer<T> serializer
-    )
-    {
+    ) {
         return executor.execute((client) -> client.selectRows(request, serializer));
     }
 
     @Override
     public <T> CompletableFuture<Void> selectRows(SelectRowsRequest request, YTreeRowSerializer<T> serializer,
-                                                  ConsumerSource<T> consumer)
-    {
+                                                  ConsumerSource<T> consumer) {
         return executor.execute((client) -> client.selectRows(request, serializer, consumer));
     }
 
@@ -369,4 +370,112 @@ public class MultiYTsaurusClient implements ImmutableTransactionalClient, Closea
             return this;
         }
     }
+}
+
+class MultiExecutor implements Closeable {
+
+    private final List<ClientEntry> clients;
+    private final Duration banPenalty;
+    private final Duration banDuration;
+    private final PenaltyProvider penaltyProvider;
+    private final MultiExecutorMonitoring executorMonitoring;
+
+    public MultiExecutor(
+            List<MultiYTsaurusClient.YTsaurusClientOptions> clientOptions,
+            Duration banPenalty,
+            Duration banDuration,
+            PenaltyProvider penaltyProvider,
+            MultiExecutorMonitoring executorMonitoring
+    ) {
+        this.clients = clientOptions.stream().map(ClientEntry::new).toList();
+        this.banPenalty = banPenalty;
+        this.banDuration = banDuration;
+        this.penaltyProvider = penaltyProvider;
+        this.executorMonitoring = executorMonitoring;
+    }
+
+    <R> CompletableFuture<R> execute(Function<BaseYTsaurusClient, CompletableFuture<R>> callback) {
+        return new MultiExecutorRequestTask<>(
+                getEffectiveClientOptions(Instant.now()),
+                callback,
+                executorMonitoring
+        ).getFuture();
+    }
+
+    private synchronized List<MultiExecutorRequestTask.ClientEntry> getEffectiveClientOptions(Instant now) {
+        // update penalties
+        for (ClientEntry client : clients) {
+            if (client.banUntil != null && client.banUntil.compareTo(now) < 0) {
+                client.adaptivePenalty = Duration.ZERO;
+            }
+            client.externalPenalty = penaltyProvider.getPenalty(client.clientOptions.getShortClusterName());
+        }
+
+        // calculate min penalty to subtract
+        final Duration minPenalty = Collections.min(clients.stream().map(ClientEntry::getPenalty).toList());
+
+        // create immutable client entries with right effective penalties
+        return clients.stream().map((client ) -> new MultiExecutorRequestTask.ClientEntry(
+                client.clientOptions,
+                client.getPenalty().minus(minPenalty),
+                client::onCompleteRequest
+        )).toList();
+    }
+
+    @Override
+    public void close() throws IOException {
+        penaltyProvider.close();
+    }
+
+    public class ClientEntry {
+
+        private final MultiYTsaurusClient.YTsaurusClientOptions clientOptions;
+        private Duration adaptivePenalty;
+        private Duration externalPenalty;
+        @Nullable
+        private Instant banUntil;
+
+        public ClientEntry(
+                MultiYTsaurusClient.YTsaurusClientOptions clientOptions
+        ) {
+            this(clientOptions, Duration.ZERO, Duration.ZERO, null);
+        }
+
+        public ClientEntry(
+                MultiYTsaurusClient.YTsaurusClientOptions clientOptions,
+                Duration adaptivePenalty,
+                Duration externalPenalty,
+                @Nullable Instant banUntil
+        ) {
+            if (clientOptions.client.getClusters().size() != 1) {
+                throw new IllegalArgumentException("Got YTsaurusClient with more than 1 cluster");
+            }
+            this.clientOptions = clientOptions;
+            this.adaptivePenalty = adaptivePenalty;
+            this.externalPenalty = externalPenalty;
+            this.banUntil = banUntil;
+        }
+
+        MultiYTsaurusClient.YTsaurusClientOptions getClientOptions() {
+            return clientOptions;
+        }
+
+        Duration getPenalty() {
+            return clientOptions.initialPenalty.plus(adaptivePenalty).plus(externalPenalty);
+        }
+
+        public synchronized void onCompleteRequest(Boolean success) {
+            if (success) {
+                if (this.adaptivePenalty.compareTo(Duration.ZERO) > 0) {
+                    this.banUntil = null;
+                    this.adaptivePenalty = Duration.ZERO;
+                }
+            } else {
+                this.banUntil = Instant.now().plus(banDuration);
+                this.adaptivePenalty = this.adaptivePenalty.plus(banPenalty);
+            }
+        }
+
+    }
+
 }
